commit before checkout to develop
