+++ app/themes/user-admin/e2e/home/home.e2e-spec.ts
@@ -1,4 +1,4 @@
-import { UserAdminPage } from './home.po';
+import { UserAdminPage } from '../app.po';
 import { protractor, browser, element, by, WebElement } from 'protractor';
 
 describe('Home Page', function() {
@@ -9,17 +9,7 @@ describe('Home Page', function() {
     });
 
     it('should login successfully with right email and password', () => {
-        page.navigateTo();
-        let email = browser.findElement(protractor.By.name('email'));
-        let password = browser.findElement(protractor.By.name('password'));
-        let submit = browser.findElement(protractor.By.className('btn btn-success'));
-
-        email.sendKeys('foo@mail.com');
-        password.sendKeys('asdf1234');
-        
-        submit.click();
-
-        expect(page.getParagraphText("app-home h1")).toEqual('Welcome to the Erdiko User Admin');
+        page.login();
     });
 
     it('should show the User Admin nav bar and its contents', () => {
@@ -28,6 +18,7 @@ describe('Home Page', function() {
 
         expect(page.getParagraphText("nav ul > li:first-child > a")).toEqual('Home');
         expect(page.getParagraphText("nav ul > li:nth-child(2) > a")).toEqual('User List');
+        expect(page.getParagraphText("nav ul > li:nth-child(3) > a")).toEqual('Users Event Log');
         expect(page.getParagraphText("nav ul > li:last-child > a")).toEqual('Logout');
     });
 
@@ -37,17 +28,11 @@ describe('Home Page', function() {
 
         expect(page.getParagraphText("app-home ul > li:first-child > a")).toEqual('Create a User');
         expect(page.getParagraphText("app-home ul > li:nth-child(2) > a")).toEqual('User List');
+        expect(page.getParagraphText("app-home ul > li:last-child > a")).toEqual('User Event Log');
     });
 
     it('should logout when Logout link is clicked', () => {
-        
-        expect(element(by.css(".nav navbar-nav > li:last-child > a"))).toBeTruthy();
-        //Logout is clicked
-
-        let logout = browser.findElement(protractor.by.css('ul > li:last-child > a'));
-        logout.click();
-
-        expect(page.getParagraphText(".navbar-brand")).toEqual('User Admin');
+        page.logout();
     });
 });
